boat/requirements.txt
paho-mqtt==2.1.0
geopy==2.4.1